drone_analysis/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
Flask-Limiter==3.5.0
opencv-python==4.8.1.78
tensorflow==2.13.0
numpy==1.24.3
Pillow==10.0.1
requests==2.31.0
python-dotenv==1.0.0
gunicorn==21.2.0
redis==5.0.1
celery==5.3.4
scikit-image==0.21.0
scipy==1.11.1
matplotlib==3.7.2
seaborn==0.12.2
geopy==2.4.0
pytz==2023.3
python-dateutil==2.8.2
werkzeug==2.3.7
